+++ src/main/java/ecomeal/client/entity/PresetBasket.java
@@ -6,7 +6,7 @@ public class PresetBasket extends Basket {
 
 	private static final long serialVersionUID = 8016189837067534406L;
 
-	public PresetBasket(int id, String name, Integer price, String category, String category_image, List<Product> products) {
+	public PresetBasket(int id, String name, Double price, String category, String category_image, List<Product> products) {
 		super(id, name, price, category, category_image, products);
 	}
